Change some settings to usize

+++ alvr/server/src/bitrate.rs
@@ -126,7 +126,7 @@ impl BitrateManager {
             if decoder_latency > Duration::from_millis(config.max_decoder_latency_ms) {
                 self.decoder_latency_overstep_count += 1;
 
-                if self.decoder_latency_overstep_count == config.latency_overstep_frames as usize {
+                if self.decoder_latency_overstep_count == config.latency_overstep_frames {
                     self.dynamic_max_bitrate =
                         f32::min(self.bitrate_average.get_average(), self.dynamic_max_bitrate)
                             * config.latency_overstep_multiplier;
